# Reuse backing array of pb slices from vt pool