Any slices coming back from `*FromVTPool` should either be nil (in which case `cap(slice) == 0`) or has a backing array (`len(slice)==0 && cap(slice) >= 0`). Reuse the backing array to reduce allocations.

In conjunction with #17 , BenchmarkCombinedMetricsEncoding will have 0 allocs/op.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/reuse-backing-array-before │ bench-out/reuse-backing-array-after │
                            │                sec/op                │    sec/op      vs base              │
AggregateCombinedMetrics-16                            3.451µ ± 3%    3.420µ ±  2%       ~ (p=0.818 n=6)
AggregateBatchSerial-16                                18.24µ ± 4%    18.13µ ±  4%       ~ (p=0.818 n=6)
AggregateBatchParallel-16                              19.46µ ± 6%    19.65µ ±  3%       ~ (p=0.589 n=6)
CombinedMetricsEncoding-16                             8.353µ ± 5%    7.612µ ±  2%  -8.88% (p=0.002 n=6)
CombinedMetricsDecoding-16                             4.239µ ± 1%    4.300µ ± 31%  +1.45% (p=0.026 n=6)
CombinedMetricsToBatch-16                              421.9µ ± 3%    417.4µ ±  2%       ~ (p=0.394 n=6)
EventToCombinedMetrics-16                              1.016µ ± 5%    1.012µ ±  4%       ~ (p=0.173 n=6)
geomean                                                10.93µ         10.77µ        -1.40%

                            │ bench-out/reuse-backing-array-before │ bench-out/reuse-backing-array-after │
                            │                 B/op                 │     B/op      vs base               │
AggregateCombinedMetrics-16                           3.626Ki ± 5%   3.655Ki ± 0%        ~ (p=0.093 n=6)
AggregateBatchSerial-16                               22.76Ki ± 0%   22.53Ki ± 1%   -1.01% (p=0.002 n=6)
AggregateBatchParallel-16                             22.63Ki ± 1%   22.63Ki ± 1%        ~ (p=0.818 n=6)
CombinedMetricsEncoding-16                            2.562Ki ± 0%   2.250Ki ± 0%  -12.20% (p=0.002 n=6)
CombinedMetricsDecoding-16                            10.01Ki ± 0%   10.01Ki ± 0%        ~ (p=1.000 n=6)
geomean                                               8.632Ki        8.406Ki        -2.62%

                            │ bench-out/reuse-backing-array-before │ bench-out/reuse-backing-array-after │
                            │              allocs/op               │ allocs/op   vs base                 │
AggregateCombinedMetrics-16                             41.00 ± 2%   37.00 ± 0%   -9.76% (p=0.002 n=6)
AggregateBatchSerial-16                                 183.0 ± 1%   169.5 ± 1%   -7.38% (p=0.002 n=6)
AggregateBatchParallel-16                               183.0 ± 0%   171.0 ± 0%   -6.56% (p=0.002 n=6)
CombinedMetricsEncoding-16                              77.00 ± 0%   37.00 ± 0%  -51.95% (p=0.002 n=6)
CombinedMetricsDecoding-16                              40.00 ± 0%   40.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                 84.19        69.20       -17.80%
¹ all samples are equal
```