@carsonip 2023-07-12
Updated PR to use `golang.org/x/exp/slices`.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/reuse-backing-array-before-slices-grow │ bench-out/reuse-backing-array-after-slices-grow │
                            │                      sec/op                      │          sec/op           vs base               │
AggregateCombinedMetrics-16                                       4.001µ ± 99%               3.666µ ±  5%   -8.39% (p=0.002 n=6)
AggregateBatchSerial-16                                           20.95µ ± 12%               19.59µ ± 26%        ~ (p=0.240 n=6)
AggregateBatchParallel-16                                         24.43µ ± 17%               28.28µ ± 25%        ~ (p=0.093 n=6)
CombinedMetricsEncoding-16                                        6.645µ ±  7%               8.218µ ±  2%  +23.68% (p=0.002 n=6)
CombinedMetricsDecoding-16                                        5.242µ ±  6%               4.524µ ±  2%  -13.69% (p=0.002 n=6)
CombinedMetricsToBatch-16                                         454.4µ ±  5%               446.2µ ±  5%        ~ (p=0.310 n=6)
EventToCombinedMetrics-16                                         1.344µ ±  6%               1.089µ ±  2%  -18.97% (p=0.002 n=6)
geomean                                                           12.34µ                     12.04µ         -2.42%

                            │ bench-out/reuse-backing-array-before-slices-grow │ bench-out/reuse-backing-array-after-slices-grow │
                            │                       B/op                       │          B/op            vs base                │
AggregateCombinedMetrics-16                                      3.331Ki ± 14%              3.578Ki ± 3%    +7.40% (p=0.002 n=6)
AggregateBatchSerial-16                                          22.11Ki ±  1%              22.69Ki ± 1%    +2.63% (p=0.002 n=6)
AggregateBatchParallel-16                                        22.13Ki ±  0%              22.69Ki ± 0%    +2.54% (p=0.002 n=6)
CombinedMetricsEncoding-16                                         576.0 ±  0%               2304.0 ± 0%  +300.00% (p=0.002 n=6)
CombinedMetricsDecoding-16                                       10.01Ki ±  0%              10.01Ki ± 0%         ~ (p=0.697 n=6)
geomean                                                          6.202Ki                    8.387Ki        +35.22%

                            │ bench-out/reuse-backing-array-before-slices-grow │ bench-out/reuse-backing-array-after-slices-grow │
                            │                    allocs/op                     │       allocs/op         vs base                 │
AggregateCombinedMetrics-16                                         34.00 ± 9%               37.00 ± 3%   +8.82% (p=0.002 n=6)
AggregateBatchSerial-16                                             159.0 ± 0%               171.0 ± 0%   +7.55% (p=0.002 n=6)
AggregateBatchParallel-16                                           159.0 ± 0%               171.0 ± 0%   +7.55% (p=0.002 n=6)
CombinedMetricsEncoding-16                                          45.00 ± 0%               37.00 ± 0%  -17.78% (p=0.002 n=6)
CombinedMetricsDecoding-16                                          40.00 ± 0%               40.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                             68.85                    69.33        +0.69%
¹ all samples are equal

```

CombinedMetricsEncoding B/op is higher and sec/op increased. The same for Go 1.21 "growSlice" is used instead of `slices.Grow`.