@carsonip 2023-07-13
Oh, I said "Updated PR" but I forgot to push the changes. The benchmark result still stands. I'll merge the current PR without `slices.Grow`. We can look into it later again.